+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ---
+## [1.1.4]
+### Added
+- used scale animation is selected Vector2 or float
+### Fixed
+- deleted unused field
+---
 ## [1.1.3]
 ### Fixed
 - can`t stopped animation fixed
@@ -204,3 +210,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.1.2]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v1.1.2
 
 [1.1.3]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v1.1.3
+
+[1.1.4]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v1.1.4